hoist 'latest_beacon_entry' from ChainStore to ChainIndex
