+++ src/chain/store/index.rs
@@ -11,6 +11,7 @@ use itertools::Itertools;
 use lru::LruCache;
 use nonzero_ext::nonzero;
 use parking_lot::Mutex;
+use crate::beacon::{BeaconEntry, IGNORE_DRAND_VAR};
 
 use crate::chain::Error;
 
@@ -144,6 +145,46 @@ impl<DB: Blockstore> ChainIndex<DB> {
                 child
             })
         })
+    }
+
+    /// Finds the latest beacon entry given a tipset up to 20 tipsets behind
+    pub fn latest_beacon_entry(&self, ts: &Tipset) -> Result<BeaconEntry, Error> {
+        let check_for_beacon_entry = |ts: &Tipset| {
+            let cbe = ts.min_ticket_block().beacon_entries();
+            if let Some(entry) = cbe.last() {
+                return Ok(Some(entry.clone()));
+            }
+            if ts.epoch() == 0 {
+                return Err(Error::Other(
+                    "made it back to genesis block without finding beacon entry".to_owned(),
+                ));
+            }
+            Ok(None)
+        };
+
+        if let Some(entry) = check_for_beacon_entry(ts)? {
+            return Ok(entry);
+        }
+        let mut cur = self.load_tipset(ts.parents())?;
+        for i in 1..20 {
+            if i != 1 {
+                cur = self.load_tipset(cur.parents())?;
+            }
+            if let Some(entry) = check_for_beacon_entry(&cur)? {
+                return Ok(entry);
+            }
+        }
+
+        if std::env::var(IGNORE_DRAND_VAR) == Ok("1".to_owned()) {
+            return Ok(BeaconEntry::new(
+                0,
+                vec![9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9],
+            ));
+        }
+
+        Err(Error::Other(
+            "Found no beacon entries in the 20 latest tipsets".to_owned(),
+        ))
     }
 }
